+++ aws/glue/prediction.py
@@ -5,15 +5,17 @@ from pyspark.context import SparkContext
 from awsglue.context import GlueContext
 from awsglue.job import Job
 
-from features.joined_with_churn import spark_job
+from features.prediction import spark_job
+from features.utils import yesterday_dt
 
 args = getResolvedOptions(sys.argv, [
     'JOB_NAME',
     'dt',
     'input_subscription',
-    'input_performance',
-    'input_activity',
+    'input_prediction',
     'output_path'])
+if args['dt'] == 'yesterday':
+    args['dt'] = yesterday_dt()
 
 sc = SparkContext()
 glueContext = GlueContext(sc)
