+++ aws/lambda/create_transform_job.py
@@ -2,20 +2,20 @@ import boto3
 import os
 from datetime import datetime, timedelta
 
-S3_INPUT_PATH = 's3://datascience-feature-eng-test-201901/users-joined-201901/'
-S3_OUTPUT_PATH = 's3://datascience-feature-eng-test-201901/sagemaker/test-output/'
-
 
 def lambda_handler(event, context):
     sagemaker = boto3.client('sagemaker')
 
     transform_inputs = event['TransformArguments']
-    if transform_inputs['TransformInput']['DataSource']['S3DataSource']['S3Uri'] == 's3://set-by-lambda':
-        yesterday_dt = (datetime.now() - timedelta(1)).strftime('%Y-%m-%d')
-        transform_inputs['TransformInput']['DataSource']['S3DataSource']['S3Uri'] = os.path.join(
-            S3_INPUT_PATH, 'dt={}'.format(yesterday_dt))
-        transform_inputs['TransformOutput']['S3OutputPath'] = os.path.join(
-            S3_OUTPUT_PATH, 'dt={}'.format(yesterday_dt))
+    yesterday_dt = (datetime.now() - timedelta(1)).strftime('%Y-%m-%d')
+
+    input_path = transform_inputs['TransformInput']['DataSource']['S3DataSource']['S3Uri']
+    transform_inputs['TransformInput']['DataSource']['S3DataSource']['S3Uri'] = os.path.join(
+        input_path, 'dt={}'.format(yesterday_dt))
+
+    transform_inputs['TransformOutput']['S3OutputPath']
+    transform_inputs['TransformOutput']['S3OutputPath'] = os.path.join(output_path, 'dt={}'.format(yesterday_dt))
+
     transform_inputs['TransformJobName'] = 'churn-daily-scoring-{}'.format(datetime.now().strftime('%Y-%m-%d-%H-%M-%S'))
     sagemaker.create_transform_job(**transform_inputs)
     return transform_inputs
