+++ aws/lambda/describe_transform_job.py
@@ -1,10 +1,9 @@
 import boto3
 import time
 
-sagemaker = boto3.client('sagemaker')
-
-
 def lambda_handler(event, context):
+    sagemaker = boto3.client('sagemaker')
+    
     job_run_state = 'InProgress'
     max_iters = 89
     current_iter = 0
@@ -13,5 +12,4 @@ def lambda_handler(event, context):
         job_run_state = sagemaker.describe_transform_job(
             TransformJobName=event['TransformArguments']['TransformJobName'])['TransformJobStatus']
         current_iter = current_iter + 1
-
     return job_run_state
