+++ aws/step_functions/etl_flow.json
@@ -19,7 +19,7 @@
     },
     "Get ETL Job Status": {
       "Type": "Task",
-      "Resource": "arn:aws:lambda:us-east-1:602761496211:function:datascience-get-job-run",
+      "Resource": "arn:aws:lambda:us-east-1:602761496211:function:datascience-glue-get-job-run",
       "Next": "ETL Job Complete?",
       "ResultPath": "$.JobRunState",
       "Retry": [
