+++ aws/step_functions/etl_input.json
