--- aws/step_functions/sample_input.json
@@ -1,27 +1,0 @@
-{
-  "JobName": "user-churn",
-  "JobArguments": {},
-  "wait_time": 1,
-  "TransformArguments": {
-    "TransformJobName": "SET-BY-LAMBDA",
-    "ModelName": "linear-learner-2019-01-23-15-08-31-382",
-    "TransformInput": {
-      "DataSource": {
-        "S3DataSource": {
-          "S3DataType": "S3Prefix",
-          "S3Uri": "s3://set-by-lambda"
-        }
-      },
-      "ContentType": "text/csv",
-      "CompressionType": "Gzip",
-      "SplitType": "Line"
-    },
-    "TransformOutput": {
-      "S3OutputPath": "s3://set-by-lambda"
-    },
-    "TransformResources": {
-      "InstanceType": "ml.c4.xlarge",
-      "InstanceCount": 1
-    }
-  }
-}
